refactor: place animation styles to corresponding components
